--- frontend/src/assets/styles/global.css
@@ -1,202 +1,0 @@
-/* Variables globales */
-:root {
-  --primary-color: #42b983;
-  --secondary-color: #2c3e50;
-  --danger-color: #ff4444;
-  --warning-color: #ffd700;
-  --success-color: #90EE90;
-  --background-color: #f8f9fa;
-  --text-color: #2c3e50;
-  --border-color: #ddd;
-  --shadow: 0 2px 4px rgba(0,0,0,0.1);
-  --border-radius: 8px;
-  --spacing-small: 0.5rem;
-  --spacing-medium: 1rem;
-  --spacing-large: 2rem;
-}
-
-/* Reset et styles de base */
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: Arial, sans-serif;
-  color: var(--text-color);
-  line-height: 1.6;
-  background-color: var(--background-color);
-}
-
-/* Layout commun */
-.container {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: var(--spacing-large);
-}
-
-/* Headings */
-h1, h2, h3, h4, h5, h6 {
-  color: var(--secondary-color);
-  margin-bottom: var(--spacing-medium);
-}
-
-h1 { font-size: 2rem; }
-h2 { font-size: 1.5rem; }
-h3 { font-size: 1.25rem; }
-
-/* Boutons */
-.btn {
-  padding: 0.8rem 1.5rem;
-  border: none;
-  border-radius: var(--border-radius);
-  cursor: pointer;
-  font-size: 1rem;
-  transition: all 0.3s ease;
-}
-
-.btn-primary {
-  background-color: var(--primary-color);
-  color: white;
-}
-
-.btn-secondary {
-  background-color: var(--secondary-color);
-  color: white;
-}
-
-.btn-danger {
-  background-color: var(--danger-color);
-  color: white;
-}
-
-.btn:hover {
-  opacity: 0.9;
-  transform: translateY(-1px);
-}
-
-/* Cards */
-.card {
-  background: white;
-  border-radius: var(--border-radius);
-  padding: var(--spacing-large);
-  box-shadow: var(--shadow);
-  margin-bottom: var(--spacing-medium);
-}
-
-/* Forms */
-.form-group {
-  margin-bottom: var(--spacing-medium);
-}
-
-.form-control {
-  width: 100%;
-  padding: 0.8rem;
-  border: 1px solid var(--border-color);
-  border-radius: var(--border-radius);
-  font-size: 1rem;
-}
-
-.form-control:focus {
-  outline: none;
-  border-color: var(--primary-color);
-}
-
-/* Grids */
-.grid {
-  display: grid;
-  gap: var(--spacing-medium);
-}
-
-.grid-2 {
-  grid-template-columns: repeat(2, 1fr);
-}
-
-.grid-3 {
-  grid-template-columns: repeat(3, 1fr);
-}
-
-@media (max-width: 768px) {
-  .grid-2, .grid-3 {
-    grid-template-columns: 1fr;
-  }
-}
-
-/* Status badges */
-.badge {
-  padding: 0.25rem 0.5rem;
-  border-radius: 4px;
-  font-size: 0.875rem;
-}
-
-.badge-success {
-  background-color: var(--success-color);
-  color: var(--text-color);
-}
-
-.badge-warning {
-  background-color: var(--warning-color);
-  color: var(--text-color);
-}
-
-.badge-danger {
-  background-color: var(--danger-color);
-  color: white;
-}
-
-/* Tables */
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-bottom: var(--spacing-large);
-}
-
-.table th, 
-.table td {
-  padding: var(--spacing-medium);
-  text-align: left;
-  border-bottom: 1px solid var(--border-color);
-}
-
-.table th {
-  background-color: var(--background-color);
-  font-weight: bold;
-}
-
-/* Loading states */
-.loading {
-  text-align: center;
-  padding: var(--spacing-large);
-  color: var(--text-color);
-}
-
-/* Empty states */
-.empty-state {
-  text-align: center;
-  padding: var(--spacing-large);
-  color: var(--text-color);
-}
-
-/* Navigation */
-.nav-container {
-  background-color: white;
-  box-shadow: var(--shadow);
-  padding: var(--spacing-medium);
-}
-
-.nav-link {
-  color: var(--text-color);
-  text-decoration: none;
-  padding: var(--spacing-small) var(--spacing-medium);
-  border-radius: var(--border-radius);
-}
-
-.nav-link:hover {
-  background-color: var(--background-color);
-}
-
-.nav-link.active {
-  color: var(--primary-color);
-  font-weight: bold;
-}
