+++ frontend/src/router/index.ts
@@ -1,12 +1,12 @@
 import { createRouter, createWebHistory } from 'vue-router';
 import Home from '../views/HomeView.vue';
-import AdminView from '../views/admin/AdminView.vue';
+import AdminView from '../views/AdminView.vue';
 import ConnectView from '../views/ConnectView.vue';
 import RegisterView from '../views/RegisterView.vue';
-import UserManagement from '../views/admin/UserManagement.vue';
-import MenuManagement from '../views/admin/MenuManagement.vue';
-import OrderManagement from '../views/admin/OrderManagement.vue';
-import StatisticsManagement from '../views/admin/StatisticsManagement.vue';
+import UserManagement from '../views/UserManagement.vue';
+import MenuManagement from '../views/MenuManagement.vue';
+import OrderManagement from '../views/OrderManagement.vue';
+import StatisticsManagement from '../views/StatisticsManagement.vue';
 import UserView from '../views/UserView.vue';
 import FeedbackView from '../views/FeedbackView.vue';
 import ModifyView from '../views/ModifyView.vue';
@@ -15,12 +15,6 @@ import CommanderView from '../views/CommanderView.vue';
 import ChefView from '../views/ChefView.vue';
 import OrderList from '../components/chef/OrderList.vue';
 import RecipeManager from '../components/chef/RecipeManager.vue';
-import ServerView from '../views/server/ServerView.vue';
-import CurrentOrders from '../components/server/CurrentOrders.vue';
-import Reservations from '../components/server/Reservations.vue';
-import ServerOrdersView from '../views/server/ServerOrdersView.vue';
-import ServerReservationsView from '../views/server/ServerReservationsView.vue';
-import ServerNewOrderView from '../views/server/ServerNewOrderView.vue';
 import ReservationView from '../views/ReservationView.vue'
 import MenuView from '../views/MenuView.vue';  
 
@@ -43,12 +37,7 @@ const routes = [
   {path: '/chef/',name: 'chef',component: ChefView, meta: { requiresAuth: true, role: 'CUISINIER' }},
   {path: '/chef/orders',name: 'ChefOrders',component: OrderList, meta: { requiresAuth: true, role: 'CUISINIER' }},
   {path: '/chef/recipes',name: 'chefRecipe',component: RecipeManager, meta: { requiresAuth: true, role: 'CUISINIER' }},
-  {path: '/server', name: 'server',component: ServerView, meta: { requiresAuth: true, role: 'SERVEUR' }},
-  {path: '/server/orders', name: 'server-orders', component: CurrentOrders, meta: { requiresAuth: true, role: 'SERVEUR' } },
-  {path: '/server/reservations', name: 'server-reservations', component: Reservations,meta: { requiresAuth: true, role: 'SERVEUR' }},
-  {path: '/server/new-order', name: 'server-new-order',component: ServerNewOrderView, meta: { requiresAuth: true, role: 'SERVEUR' }},
-  {path: '/user/reservation',name: 'reservation',component: ReservationView ,meta: { requiresAuth: true, role: 'CLIENT' }}
-
+  {path: '/user/reservation',name: 'reservation',component: ReservationView ,meta: { requiresAuth: true, role: 'CLIENT' }},
 ];
 
 const router = createRouter({
