+++ frontend/src/views/AdminView.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script setup>
-import NavBar from '../../components/NavBar.vue';
+import NavBar from '../components/NavBar.vue';
 </script>
 
 <style scoped>
