+++ frontend/src/views/MenuManagement.vue
@@ -28,7 +28,7 @@
 <script setup >
 import { ref } from 'vue'
 import axios from 'axios'
-import NavBar from '../../components/NavBar.vue';
+import NavBar from '../components/NavBar.vue';
 
 const dishes = ref([])
 const currentDish = ref({ id: null, name: '', price: 0, allergens: '', category: 'entrée', stock: 0 })
