+++ frontend/src/views/OrderManagement.vue
@@ -40,7 +40,7 @@
 <script setup >
 import { ref } from 'vue'
 import axios from 'axios'
-import NavBar from '../../components/NavBar.vue';
+import NavBar from '../components/NavBar.vue';
 
 const orders = ref([])
 const dishes = ref([])
